requirements.txt
streamlit==1.28.1
mysql-connector-python==8.0.33
pandas==2.0.3
requests==2.31.0
numpy==1.24.3
